server/requirements.txt
uvicorn
fastapi
python-dotenv
pytest
requests
aiormq
termcolor
asyncio
pika
aio_pika